aider: Make the Supernova effect fullscreen.
# Aider chat conversation:

USER: Please update the Supernova effect in index.tsx to be fullscreen. The entire screen should be black and show the logo effect.

ASSISTANT: To make the Supernova effect fullscreen, we need to modify the `pages/index.tsx` file. We will add a new CSS class to the section that contains the SupernovaLogo component and the Particles component. This class will make the section take up the full screen.

Here is the change:

```tsx
pages/index.tsx
<<<<<<< HEAD
      <section className="bg-traidmark-black dark:bg-gray-900">
        <Particles />
        <SupernovaLogo />
      </section>
=======
      <section className="bg-traidmark-black dark:bg-gray-900 w-screen h-screen flex items-center justify-center">
        <Particles />
        <SupernovaLogo />
      </section>
>>>>>>> updated
```

In the updated code, `w-screen` and `h-screen` are Tailwind CSS classes that set the width and height of the section to the width and height of the screen, respectively. The `flex`, `items-center`, and `justify-center` classes are used to center the SupernovaLogo component both vertically and horizontally within the section.