aider: Update the SupernovaLogo to use the Inter font and fill at least 30% of the screen.
# Aider chat conversation:

USER: Update the SupernovaLogo to be inter font and fill at least 30% of the screen

ASSISTANT: To make the "trAIdmark" text use the Inter font and fill at least 30% of the screen, we need to modify the `components/SupernovaLogo.tsx` file. We will add CSS classes to the `text` element to set the font and size.

Here is the change:

```tsx
components/SupernovaLogo.tsx
<<<<<<< HEAD
    <svg ref={logoRef} className="logo">
      <text x="0" y="45" fill="white">trAIdmark</text>
    </svg>
=======
    <svg ref={logoRef} className="logo">
      <text x="0" y="45" fill="white" className="font-inter text-9xl">trAIdmark</text>
    </svg>
>>>>>>> updated
```

In the updated code, `font-inter` is a Tailwind CSS class that sets the font to Inter. The `text-9xl` class sets the font size to a large value that should fill at least 30% of the screen. You might need to adjust this value depending on the exact size of the screen and the aspect ratio of the "trAIdmark" text.

Please note that this change assumes that you have configured Tailwind CSS to include the Inter font and the `text-9xl` size. If you haven't, you will need to add them to your Tailwind CSS configuration.