aider: Add CSS animation to change the background color of the main section.
# Aider chat conversation:

ASSISTANT: To achieve this, we will need to modify the `pages/index.tsx` file to add a CSS animation that changes the background color. We will also need to add the keyframes for this animation in the `styles/tailwind.css` file.

Here are the changes:

```tsx
pages/index.tsx
<<<<<<< HEAD
<section className="bg-traidmark-black dark:bg-gray-900 w-screen h-screen flex items-center justify-center relative">
=======
<section className="bg-oscillate w-screen h-screen flex items-center justify-center relative">
>>>>>>> updated
```

```css
styles/tailwind.css
<<<<<<< HEAD
=======
@keyframes oscillate {
  0% { background-color: #000000; }
  33% { background-color: #0000FF; }
  66% { background-color: #FFD700; }
  100% { background-color: #000000; }
}

.bg-oscillate {
  animation: oscillate 10s infinite;
}
>>>>>>> updated
```

These changes will add a CSS animation that changes the background color of the main section from black to blue to mustard and back to black. The animation will repeat indefinitely.